// src/lib/sdk/blockfrost/api.ts
export interface TxContent {
    /**
     * Transaction hash
     * @type {string}
     * @memberof TxContent
     */
    'hash': string;
    /**
     * Block hash
     * @type {string}
     * @memberof TxContent
     */
    'block': string;
    /**
     * Block number
     * @type {number}
     * @memberof TxContent
     */
    'block_height': number;
    /**
     * Block creation time in UNIX time
     * @type {number}
     * @memberof TxContent
     */
    'block_time': number;
    /**
     * Slot number
     * @type {number}
     * @memberof TxContent
     */
    'slot': number;
    /**
     * Transaction index within the block
     * @type {number}
     * @memberof TxContent
     */
    'index': number;
    /**
     * 
     * @type {Array<TxContentOutputAmountInner>}
     * @memberof TxContent
     */
    'output_amount': Array<TxContentOutputAmountInner>;
    /**
     * Fees of the transaction in Lovelaces
     * @type {string}
     * @memberof TxContent
     */
    'fees': string;
    /**
     * Deposit within the transaction in Lovelaces
     * @type {string}
     * @memberof TxContent
     */
    'deposit': string;
    /**
     * Size of the transaction in Bytes
     * @type {number}
     * @memberof TxContent
     */
    'size': number;
    /**
     * Left (included) endpoint of the timelock validity intervals
     * @type {string}
     * @memberof TxContent
     */
    'invalid_before': string | null;
    /**
     * Right (excluded) endpoint of the timelock validity intervals
     * @type {string}
     * @memberof TxContent
     */
    'invalid_hereafter': string | null;
    /**
     * Count of UTXOs within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'utxo_count': number;
    /**
     * Count of the withdrawals within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'withdrawal_count': number;
    /**
     * Count of the MIR certificates within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'mir_cert_count': number;
    /**
     * Count of the delegations within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'delegation_count': number;
    /**
     * Count of the stake keys (de)registration within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'stake_cert_count': number;
    /**
     * Count of the stake pool registration and update certificates within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'pool_update_count': number;
    /**
     * Count of the stake pool retirement certificates within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'pool_retire_count': number;
    /**
     * Count of asset mints and burns within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'asset_mint_or_burn_count': number;
    /**
     * Count of redeemers within the transaction
     * @type {number}
     * @memberof TxContent
     */
    'redeemer_count': number;
    /**
     * True if contract script passed validation
     * @type {boolean}
     * @memberof TxContent
     */
    'valid_contract': boolean;
}
// ...
export interface TxContentOutputAmountInner {
    /**
     * The unit of the value
     * @type {string}
     * @memberof TxContentOutputAmountInner
     */
    'unit': string;
    /**
     * The quantity of the unit
     * @type {string}
     * @memberof TxContentOutputAmountInner
     */
    'quantity': string;
}
// ...
export interface TxContentRedeemersInner {
    /**
     * Index of the redeemer within the transaction
     * @type {number}
     * @memberof TxContentRedeemersInner
     */
    'tx_index': number;
    /**
     * Validation purpose
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'purpose': TxContentRedeemersInnerPurposeEnum;
    /**
     * Script hash
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'script_hash': string;
    /**
     * Redeemer data hash
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'redeemer_data_hash': string;
    /**
     * Datum hash
     * @type {string}
     * @memberof TxContentRedeemersInner
     * @deprecated
     */
    'datum_hash': string;
    /**
     * The budget in Memory to run a script
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'unit_mem': string;
    /**
     * The budget in CPU steps to run a script
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'unit_steps': string;
    /**
     * The fee consumed to run the script
     * @type {string}
     * @memberof TxContentRedeemersInner
     */
    'fee': string;
}

export const TxContentRedeemersInnerPurposeEnum = {
    Spend: 'spend',
    Mint: 'mint',
    Cert: 'cert',
    Reward: 'reward'
} as const;

export type TxContentRedeemersInnerPurposeEnum = typeof TxContentRedeemersInnerPurposeEnum[keyof typeof TxContentRedeemersInnerPurposeEnum];
// ...
export interface TxContentUtxo {
    /**
     * Transaction hash
     * @type {string}
     * @memberof TxContentUtxo
     */
    'hash': string;
    /**
     * 
     * @type {Array<TxContentUtxoInputsInner>}
     * @memberof TxContentUtxo
     */
    'inputs': Array<TxContentUtxoInputsInner>;
    /**
     * 
     * @type {Array<TxContentUtxoOutputsInner>}
     * @memberof TxContentUtxo
     */
    'outputs': Array<TxContentUtxoOutputsInner>;
}
/**
 * 
 * @export
 * @interface TxContentUtxoInputsInner
 */
export interface TxContentUtxoInputsInner {
    /**
     * Input address
     * @type {string}
     * @memberof TxContentUtxoInputsInner
     */
    'address': string;
    /**
     * 
     * @type {Array<TxContentOutputAmountInner>}
     * @memberof TxContentUtxoInputsInner
     */
    'amount': Array<TxContentOutputAmountInner>;
    /**
     * Hash of the UTXO transaction
     * @type {string}
     * @memberof TxContentUtxoInputsInner
     */
    'tx_hash': string;
    /**
     * UTXO index in the transaction
     * @type {number}
     * @memberof TxContentUtxoInputsInner
     */
    'output_index': number;
    /**
     * The hash of the transaction output datum
     * @type {string}
     * @memberof TxContentUtxoInputsInner
     */
    'data_hash': string | null;
    /**
     * CBOR encoded inline datum
     * @type {string}
     * @memberof TxContentUtxoInputsInner
     */
    'inline_datum': string | null;
    /**
     * The hash of the reference script of the input
     * @type {string}
     * @memberof TxContentUtxoInputsInner
     */
    'reference_script_hash': string | null;
    /**
     * Whether the input is a collateral consumed on script validation failure
     * @type {boolean}
     * @memberof TxContentUtxoInputsInner
     */
    'collateral': boolean;
    /**
     * Whether the input is a reference transaction input
     * @type {boolean}
     * @memberof TxContentUtxoInputsInner
     */
    'reference'?: boolean;
}
/**
 * 
 * @export
 * @interface TxContentUtxoOutputsInner
 */
export interface TxContentUtxoOutputsInner {
    /**
     * Output address
     * @type {string}
     * @memberof TxContentUtxoOutputsInner
     */
    'address': string;
    /**
     * 
     * @type {Array<TxContentOutputAmountInner>}
     * @memberof TxContentUtxoOutputsInner
     */
    'amount': Array<TxContentOutputAmountInner>;
    /**
     * UTXO index in the transaction
     * @type {number}
     * @memberof TxContentUtxoOutputsInner
     */
    'output_index': number;
    /**
     * The hash of the transaction output datum
     * @type {string}
     * @memberof TxContentUtxoOutputsInner
     */
    'data_hash': string | null;
    /**
     * CBOR encoded inline datum
     * @type {string}
     * @memberof TxContentUtxoOutputsInner
     */
    'inline_datum': string | null;
    /**
     * Whether the output is a collateral output
     * @type {boolean}
     * @memberof TxContentUtxoOutputsInner
     */
    'collateral': boolean;
    /**
     * The hash of the reference script of the output
     * @type {string}
     * @memberof TxContentUtxoOutputsInner
     */
    'reference_script_hash': string | null;
    /**
     * Transaction hash that consumed the UTXO or null for unconsumed UTXOs. Always null for collateral outputs.
     * @type {string}
     * @memberof TxContentUtxoOutputsInner
     */
    'consumed_by_tx'?: string | null;
}